Devhuym29 (#181)
优惠券

+++ common/src/main/java/com/yumu/hexie/model/agent/AgentRepository.java
@@ -10,7 +10,7 @@ public interface AgentRepository extends JpaRepository<Agent, Long>{
 	@Query(value = "select a.id from agent a where a.status =?1 "
 			+ "and IF (?2!='', a.agentNo = ?2, 1=1 )"
 			+ "and IF (?3!='', a.name like CONCAT('%',?3,'%'), 1=1) ", nativeQuery = true)
-	List<Integer> findByAgentNoOrName(int status, String agentNo, String agentName);
+	List<Long> findByAgentNoOrName(int status, String agentNo, String agentName);
 	
 	Agent findByAgentNo(String agentNo);
 }
